Add cart to the DOM
Once we have added our product to our data, to our cart array, now of course would also want to display it when we're opening up the cart. But before we open up the cart, we would want to add it to the DOM so that way we can clearly see that we have added the product to the DOM and we do that in a separate function here just so we don't overpopulate the setupCart.js file.