Setup the update items functionality
What we can do now is find that specific one item that we want to update when we add it in the cart, and we'll update them. Now not only we can add new item if the item is not in cart, but also if the item is in the cart then we update the values. And then regardless of which action we perform, we display correct items, we display total, we set up the storage and we also open up the cart so that the user can see what is happening.