Fix a bug with the products rendering in the "Products" page and fix another tiny bug with the removeItem parameter in setupCart.js
About the first bug - if the local storage was empty and we clicked on "Products", we had no rendered product on the screen because we set up that store once the user navigates to the next page.
How do we fix that - we created a unit function, which is now running once the page loads. That way, we are just nicely grouping all of our logic. We set up a conditional fetch where if we don't have anything in the local storage, then we set up the fetch request just like we have in the index page.