Fix bug with the added product quantity
If we added the first two elements and then applied some filters and when we added a third product after applied the filters - its quantity wasn't 1 because multiple event listeners were already added/activated.
We fix this by adding one more parameter to the function. So now not only we're looking for products and element, we also check if it is filters that are calling this function and if that is the case, then we'll simply return from the function because we know that if we're calling filters, we already wokring in the products page.
We also added true to all displays functions.